==> R/01_load.qmd <==
---
title: "01_load"
author: "Group 29: Helena Møller, Emma Christensen, Frederik Holmstrup, Casper Harreby & Laura Christiansen"
format: html
---

## Introduction

The data used for this project is the data used for the article "The International Stroke Trial (IST): a randomised trial of aspirin, subcutaneous heparin, both, or neither among 19435 patient with acute ischaemic stroke" by the International Stroke Trial Collaborative Group. The purpose of this project is to improve some of the figures from the article as well as trying to get further insights. The purpose of the study and our project is to access the efficacy and safety of antithrombic therapy for treatment of acute ischemic strokes. Two different doses of heparin is tested together with and without a dose of aspirin. That creates 6 groups of patients, where one of them is the control group with no treatment.

## Loading relevant libraries

```{r}
#| echo: true 
#| eval: true
#| message: false

library("tidyverse")
library("here")
```

## Loading relevant data

The data webpage link are used for a direct load of the data to ensure reproducability

```{r}
#| echo: true 
#| eval: true
#| message: false

#The data is loaded directly from the webpage
raw_data <- read_csv(file = "https://datashare.ed.ac.uk/bitstream/handle/10283/124/IST_corrected.csv?sequence=5&isAllowed=y")
```

## Saving data in a tsv format

The data is saved in a data/\_raw folder on your local directory. The .gitignore is modified to include the entire data folder.

```{r}
#| echo: true 
#| eval: true
#| message: false

# Define the file path using the here terminology
file_path <- here("data/_raw/")

#Checking if the folder path already exists in your local directory
if (!file.exists(file_path)){dir.create(file_path, recursive = TRUE)}

#Write the data to a TSV file
write_tsv(x = raw_data, 
          file = str_c(file_path, "01_dat_load.tsv"))
```
