==> R/03_augment.qmd <==
---
title: "03_augment"
author: "Group 29: Helena Møller, Emma Christensen, Frederik Holmstrup, Casper Harreby & Laura Christiansen"
format: html
execute: 
  message: FALSE
  warning: FALSE
---

# Data Augmentation

## Loading libraries

```{r}
#| eval: true
#| echo: true

library("tidyverse")
library("here")
```

## Loading clean data

```{r}
data_clean <- read_tsv(here("data/02_dat_clean.tsv"))
```

## Data augmentation

DLH14 describes if the subject had a low dose of heparin (Yes/No), DMH14 describes if the subject had a medium dose of heparin (Yes/No), given for 14 days or till death/discharge. We want to merge the two columns into one to improve tidyness, Introduce L for low, M for medium, C for control (no Heparin) and NA for no known information or conflicting information in the two columns.

```{r}
#creating groups for heparin doses
data_clean_hepgroups <- data_clean |> 
  mutate("DH14" = case_when(
    (is.na(DLH14) & is.na(DMH14)) ~ NA,
    (DLH14 == "Y" & DMH14 == "Y") ~ NA,
    (DLH14 == "N" & DMH14 == "N") ~ "C",
    (DLH14 == "Y" & DMH14 == "N") ~ "L",
    (DLH14 == "N" & DMH14 == "Y") ~ "M"))
```

There are 6 different groups in the data. We therefore want to make a group name column to get an easy overview of the groups in the data

```{r}
#creating groups based on treatment
data_clean_trtgroups <- data_clean_hepgroups |> 
  mutate("data_group" = case_when(
    (RXASP == "Y" & RXHEP == "L") ~ "Asp-Hep(low)",
    (RXASP == "Y" & RXHEP == "M") ~ "Asp-Hep(med)",
    (RXASP == "Y" & RXHEP == "N") ~ "Asp-Hep(No)",
    (RXASP == "N" & RXHEP == "L") ~ "Asp(No)-Hep(low)",
    (RXASP == "N" & RXHEP == "M") ~ "Asp(No)-Hep(med)",
    (RXASP == "N" & RXHEP == "N") ~ "Asp(No)-Hep(No)"
  ))
```

## Saving augmented data

```{r}
write_tsv(data_clean_trtgroups, file = here("data/03_dat_aug.tsv"))
```

##
