==> R/04_describe.qmd <==
---
title: "04_describe"
author: "Group 29: Helena Møller, Emma Christensen, Frederik Holmstrup, Casper Harreby & Laura Christiansen"
format: html
execute: 
  message: FALSE
  warning: FALSE
---

# Descriptive analysis

## Loading Libraries

```{r}
#| echo: true
#| eval: true

library("tidyverse") 
library("ggplot2") 
library("dplyr") 
library("readr") 
library("here")
```

## Loading data

```{r}
#| eval: true
data <- read_tsv(here("data/03_dat_aug.tsv"))
```

## Observations and (augmented) variables

```{r}

# Show number of observations and (augmented) variables in the study population

obs_and_variables <- data |> 
  dim() |> 
  tibble()

obs_and_variables |> 
  write_tsv(here("results/04_obs_and_variables.tsv"))
```

There are 19435 observations distributed in 97 (augmented) variables in the trial.

The variabes have different names shown below:

```{r}

data |>    
  colnames()
```

## Gender distribution of study population

This section shows the gender distribution of the study population:

```{r}

# Gender distribution visualized using count()
data |> 
  count(SEX)
```

In the study population, there are 9028 females and 10407 males.

## Age distribution of study population

This section explores the age distribution of the study population:

```{r}

# Calculate the mean age
mean_age <- mean(data$AGE, na.rm = TRUE)

# Excluding missing or negative ages
data <- data |> 
  filter(!is.na(AGE) & AGE >= 0)

# Visualizing the age distribution
p1 <- ggplot(data, aes(x = AGE)) +      
  geom_histogram(fill = "lightblue", 
                 color = "darkblue", 
                 bins = 30, 
                 alpha = 0.7) + 
  labs(
    title = "Age Distribution of the Study Population",             
    x = "Age (years)", 
    y = "Number of People"
  ) +      
  theme_minimal(base_size = 14) +      
  theme(
    plot.title = element_text(hjust = 0.5, face = "bold", size = 18), 
    axis.title.x = element_text(face = "bold", size = 14),            
    axis.title.y = element_text(face = "bold", size = 14),         
    panel.grid.minor = element_blank()
  ) +                
  scale_x_continuous(breaks = seq(0, 100, by = 10), 
                     limits = c(0, 100)) +     
  scale_y_continuous(expand = expansion(mult = c(0, 0.05)), 
                     breaks = seq(0, 3000, by = 500)) +    
  geom_segment(
    x = mean_age, xend = mean_age, y = 0, yend = 2900, 
    color = "darkgreen", linetype = "dashed", size = 1.0
  ) +      
  annotate(
    "label", 
    x = mean_age + 12, y = 2400, 
    label = paste("Mean Age:", round(mean_age, 0), "years"),                 
    color = "white", fill = "darkgreen", 
    fontface = "bold", size = 4,                      
    label.padding = unit(0.4, "lines")  
    # Mean age shown in a text box with padding 
  )

# Display the plot
print(p1)
```

From the histogram distributed by age, it can be seen that the mean age of the study population is 72 years.

## Age distribution of stroke subtypes (STYPE)

The stroke subtypes (STYPE) are here distributed by age and visualized with boxplots. Here's a brief description of the different subtypes:

-   **LACS (Lacunar Stroke Syndrome)**

    Blockage of small penetrating arteries in deep brain structures, such as the basal ganglia, thalamus, or internal capsule. The prognosis is genereally better than other subtypes; smaller infarct size, where infarct size refers to the portion of the brain that has suffered irreversible damage from a stroke due to insufficient oxygen and nutrients.

-   **PACS (Partial Anterior Circulation Syndrome)**

    Partial occlusion in the anterior cerebral circulation, typically affecting the middle cerebral artery (MCA) or anterior cerebral artery (ACA). The prognosis is moderate; larger infarct than LACS, but less severe than e.g TACS.

-   **POCS (Posterior Circulation Syndrome)**

    Infarction in the posterior cerebral circulation, affecting the brainstem, cerebellum, or occipital lobes. The prognosis is variable, depending on location and size of the infarction.

-   **TACS (Total Anterior Circulation Syndrome)**

    Complete occlusion of the anterior cerebral circulation, typically affecting the middle cerebral artery (MCA) or internal carotid artery (ICA). The prognosis is poor; associated with large infarcts, severe disability, and high mortality rates.

-   **OTH** refers to other types of strokes.

```{r}

p2 <- data |> 
  ggplot(aes(x = STYPE, y = AGE, fill = STYPE)) + 
  geom_boxplot(alpha = 0.7, outlier.shape = NA) + # Boxplot without outlier points
  labs(
    title = "Stroke Subtypes Distributed by Age",
    x = "Stroke Type (STYPE)",
    y = "Age (years)"
  ) + 
  theme_minimal(base_size = 14) + 
  theme(
    plot.title = element_text(hjust = 0.5, face = "bold", size = 18),
    axis.title.x = element_text(face = "bold", size = 14),
    axis.title.y = element_text(face = "bold", size = 14),
    legend.position = "right"
  )

# Display the plot
print(p2)
```

```{r}
STYPE_median <- data |>
  group_by(STYPE) |>
  summarize(Median_AGE = round(median(AGE, na.rm = TRUE), 0))

print(STYPE_median)
```

The median age of the different stroke subtypes are shown, where the most severe subtype TACS with high mortality rates, has the highest median age of 76 years.

# Data visualization

The study population has been divided into six data groups based on the dose of heparin (no, low, medium) and whether aspirin was located. Below is a summary of the number of patients in each test group:

```{r}

# Summarize and count patients by data group
data_summary <- data |> 
  group_by(data_group) |> 
  count(name = "patient_count") 

print(data_summary)
```

## Predicted probability of death by gender at 6 months

This section visualizes the predicted probability of death at six months (EXPD6) by gender. A boxplot is used to compare probabilities.

```{r}

# Create boxplot for predicted probability of death by gender
p3 <- data |> 
  ggplot(aes(x = SEX, y = EXPD6, fill = SEX)) +   
  geom_boxplot(alpha = 0.7, color = "black") +    
  scale_fill_manual(values = c("darkred", "blue")) + 
  labs(
    title = "Predicted Probability of Death at 6 Months",     
    x = "Gender",     
    y = "Predicted Probability (EXPD6)",     
    fill = "Gender"
  ) + 
  theme_minimal(base_size = 14) +   
  theme(
    plot.title = element_text(hjust = 0.5, face = "bold", size = 15),
    axis.title.x = element_text(face = "bold", size = 10), 
    axis.title.y = element_text(face = "bold", size = 10), 
    legend.position = "top"
  ) +   
  scale_y_continuous(
    limits = c(0, 1),     
    expand = expansion(mult = c(0.05, 0.1))
  )

# Display the plot
print(p3)
```

## Comparison of median of EXPD6 by gender

The following table summarizes the median predicted probability of death (EXPD6) by gender.

```{r}
#Show median EXPD6 by gender  

median_table <- data |>
  group_by(SEX) |>
  summarize(Median_EXPD6 = median(EXPD6, na.rm = TRUE)) 

print(median_table)
```

From the table, females have a higher median probability of death at 6 months (0.1888) compared to males (0.1390).

## Save the results

```{r}
# Save the plots as PNG files

ggsave(here("results/04_description_plot_1.png"), 
       plot = p1, 
       width = 8, 
       height = 4, 
       dpi = 300, 
       bg = "white")

ggsave(here("results/04_description_plot_2.png"), 
       plot = p2, 
       width = 8, 
       height = 4, 
       dpi = 300, 
       bg = "white")

ggsave(here("results/04_description_plot_3.png"), 
       plot = p3, 
       width = 8, 
       height = 4, 
       dpi = 300, 
       bg = "white")
```
