==> R/06_analysis_2.qmd <==
---
title: "06_analysis_2"
author: "Group 29: Helena Møller, Emma Christensen, Frederik Holmstrup, Casper Harreby & Laura Christiansen"
format: html
execute: 
  message: FALSE
  warning: FALSE
---

# Data modeling

This analysis investigates whether `AGE` can be used to predict the following probabilities making a simple linear model:

-   `EXPDD` (Predicted probability of death/dependence at 6 month)

-   `EXPD6` (Predicted probability of death at 6 months)

-   `EXPD14` (Predicted probability of death at 14 days)

We will visualize the relationships between `AGE` and these outcomes using scatter plots.

## Loading Libraries

```{r}
#| eval: true
#| echo: true

library("tidyverse") 
library("ggplot2") 
library("dplyr") 
library("readr") 
library("here")
```

## Import the dataset

```{r}

# Load the dataset   
data <- read_tsv(here("data/03_dat_aug.tsv"))
```

### Model 1: Predicting `EXPDD` with `AGE`

**Linear regression:**

```{r}

# Linear regression for EXPDD 
model_1 <- data |> 
  filter(STYPE == "LACS") |> 
  lm(data = _,
     EXPDD ~ AGE)
```

```{r}

# Display model summary 
model_1 |> 
  summary()
```

From the model summary, R-squared is relatively low suggesting a poor fit if the model to the data.

```{r}


# Plot AGE vs log2(EXPDD) to observe the relationship
p1 <- data |> 
  filter(STYPE == "LACS") |> 
  ggplot(aes(x = AGE, y = log2(EXPDD))) +
  geom_point(color = "purple") +
  labs(
    title = "AGE vs. Predicted Death/Dependence Probability at 6 Months",
    x = "Age (Years)",
    y = "log2(Predicted Death/Dependence Probability)",
    caption = "Relationship modeled using simple linear regression"
  ) +
  theme_minimal()

print(p1)

```

Visualizing the log-transformed **predicted probability of death/dependence** **at 6 months**, we are able to see a tendency that older patients have a higher probability of death/dependence at 6 months.

### Model 2: Predicting `EXPD6` with `AGE`

**Linear regression:**

```{r}

# Linear regression for EXPD6
model_2 <- data |> 
  filter(STYPE == "LACS") |> 
  lm(data = _,
     log2(EXPD6) ~ AGE)
```

**Model summary:**

```{r}

# Display model summary 
model_2 |> 
  summary()
```

From the model summary, R-squared is relatively low suggesting a poor fit if the model to the data. However, the fit is better than the previous (R-squared = 0.9153)

**Visualization:**

```{r}

# Plot AGE vs log2(EXPD6) to observe the relationship
p2 <- data |> 
  filter(STYPE == "LACS") |> 
  ggplot(aes(x = AGE, y = log2(EXPD6))) +
  geom_point(color = "blue") +
  labs(
    title = "AGE vs. Predicted Death Probability at 6 Months",
    x = "Age (Years)",
    y = "log2(Predicted Death Probability)",
    caption = "Relationship modeled using simple linear regression"
  ) +
  theme_minimal()

print(p2)
```

Visualizing the log-transformed **predicted probability of death** **at 6 months**, we are able to see a tendency that older patients have a higher probability of death/dependence at 6 months. This variable is better predicted that the previous.

### Model 3: Predicting `EXPD14` with `AGE`

**Linear regression:**

```{r}

# Linear regression for EXPD14
model_3 <- data |> 
  filter(STYPE == "LACS") |> 
  lm(data = _, log2(EXPD14) ~ AGE)
```

**Model summary:**

```{r}

# Display model summary
model_3 |> 
  summary()
```

From the model summary, R-squared is relatively low suggesting a poor fit if the model to the data. This fit is the worst of the three variables we are investigating.

**Visualization:**

```{r}

# Plot AGE vs log2(EXPD14) to observe the relationship
p3 <- data |> 
  filter(STYPE == "LACS") |> 
  ggplot(aes(x = AGE, y = log2(EXPD14))) +
  geom_point(color = "red") +
  labs(
    title = "AGE vs. Predicted Death Probability at 14 Days",
    x = "Age (Years)",
    y = "log2(Predicted Death Probability)",
    caption = "Relationship modeled using simple linear regression"
  ) +
  theme_minimal()

print(p3)
```

Again, visualizing the log-transformed **predicted probability of death** **at 14 days**, we are able to see a tendency that older patients have a higher probability of death/dependence at 6 months. It even looks like there groups in the data where a line would be able to separate them.

## Saving Plots

To document the results, the plots are saved in the `results` folder.

```{r}

ggsave(filename = here("results/06_keyplot_2a.png"), 
       plot = p1, 
       width = 8, 
       height = 4, 
       dpi = 300, 
       bg = "white")  
ggsave(filename = here("results/06_keyplot_2b.png"), 
       plot = p2, 
       width = 8, 
       height = 4, 
       dpi = 300, 
       bg = "white")
ggsave(filename = here("results/06_keyplot_2c.png"), 
       plot = p3, 
       width = 8, 
       height = 4, 
       dpi = 300, 
       bg = "white")
```
