==> R/07_analysis_3.qmd <==
---
title: "07_analysis_3"
author: "Group 29: Helena Møller, Emma Christensen, Frederik Holmstrup, Casper Harreby & Laura Christiansen"
format: html
---

# Data analysis

We aim to visualize the overall time it takes before each person dies no matter the cause. The aim then is to plot the 6 different treatment combinations: aspirin or no aspirin and low, medium or no heparin for different causes of death for further analysis of which treatment works best.

## Loading libraries

```{r}
#| echo: true
#| eval: true

library("tidyverse")
library("here")
library("scales")
```

## Loading data

The augmented data is loaded.

```{r}
#| echo: true
#| eval: true
#| message: false

data <- read_tsv(here("data/03_dat_aug.tsv"))
```

## Preparing data

Prior to doing the analysis the augmented data is grouped by the cause of death, which are created into a new column having the different causes of death as values.

```{r}
#| echo: true
#| eval: true
#| message: false

# Grouping the data by the variable "DEADC"
grouped_data <- data |>
  group_by(DEADC) |>
  # Creating a new column "TD_causes", which categorizes each cause of death based on "DEADC"
  mutate( 
    TD_causes = case_when( 
      DEADC == 1 ~ "cause_1",
      DEADC == 2 ~ "cause_2",
      DEADC == 3 ~ "cause_3",
      DEADC == 4 ~ "cause_4",
      DEADC == 5 ~ "cause_5",
      DEADC == 6 ~ "cause_6",
      DEADC == 7 ~ "cause_7",
      DEADC == 8 ~ "cause_8",
    )
  )

#Filter data for each cause of death
causes <- str_c("cause_", 1:8) #Creates a list of causes
group_data_list <- causes |> 
  set_names() |> 
  map(.x = ,
      .f = ~ grouped_data |>
        filter(TD_causes == .x)) # Filter the data for each cause and stores it in a list
```

## Plot the overall data count

The patients died pr. day are plotted for the first six months. It is noticed that most of the patients die within 20 days of randomization. This aim of this is to get an overall idea of how the treatment is working.

```{r}
#| echo: true
#| eval: true
#| warning: false

#Count the occurences of time of death for the whole dataset if the person is dead at the 6 month follow up
group_counts <- grouped_data |>
  filter(FDEAD == "Y") |> 
  count(TD)

#Plots the time of death (TD) vs. the number of occurences.
p0 <- group_counts |>
  ggplot(mapping = aes(x = TD,
                       y = n))+
  geom_line()+
  coord_cartesian(xlim = c(0, 180))+
  labs(
    title =  "Most patients die within the first 20 days after randomisation",
    x = "Days elapsed since randomisation",
    y = "Number of people died x days after randomization")+
  theme_minimal()+
  theme(plot.title = element_text(hjust = 0.5))

#Display til plot
p0
  
# Save the plot as a PNG
ggsave(filename = here("results/07_keyplot_3a.png"), 
       plot = p0, 
       width = 8, 
       height = 4, 
       dpi = 300, 
       bg = "white")

```

## Creating individual plots for each cause of death

For later comparison, the amount of patient dying of ischaemic stroke, and pulomonary embolism are pr day are plotted. It is included which of the 6 treatments the patients were having. The aim of this is to see which initial appears to work best.

```{r}
#| echo: true
#| eval: true
#| warning: false

#Count the occurences of "TD" for only those died of ischaemic stroke
group_1_counts <- group_data_list |>
  pluck("cause_1") |> 
  group_by(data_group, TD) |>
  summarise(count = n(), .groups = "drop")

#Plots the time of death (TD) vs. the number of occurences.
p1 <- group_1_counts |> 
  ggplot(mapping = aes(x = TD, 
                       y = count, 
                       color = data_group)) + 
  geom_line() +
  coord_cartesian(xlim = c(0, 20))+
  labs(
    title =  "Days elapsed since randomisation before death of initial stroke",
    x = "Days elapsed since randomisation",
    y = "Number of peopled died of initial stroke after x days"
  ) +
  facet_wrap(~ data_group)

# Save the plot as a PNG
ggsave(filename = here("results/07_keyplot_3b.png"), 
       plot = p1, 
       width = 8, 
       height = 4, 
       dpi = 300, 
       bg = "white")

#Display the plot
p1

#Count the occurences of "TD" for only those died of pulmonary embolism
group_6_counts <- group_data_list |>
  pluck("cause_6") |> 
  group_by(data_group, TD) |>
  summarise(count = n(), .groups = "drop")

#Plots the time of death (TD) vs. the number of occurences.
p6 <- group_6_counts |> 
  ggplot(mapping = aes(x = TD, 
                       y = count, 
                       color = data_group)) + 
  geom_line() +
  coord_cartesian(xlim = c(0, 20))+
  labs(
    title =  "Days elapsed since randomisation before death of pulmonary embolism",
    x = "Days elapsed since randomisation",
    y = "Number of peopled died of pulmonary embolism after x days"
  ) +
  facet_wrap(~ data_group)

#Display til plot
p6

# Save the plot as a PNG
ggsave(filename = here("results/07_keyplot_3c.png"), 
       plot = p6, 
       width = 8, 
       height = 4, 
       dpi = 300, 
       bg = "white")
```
